Refactored af

+++ Problem040.cs
@@ -1,5 +1,3 @@
-using System.Text;
-
 namespace Project_Euler;
 
 public class Problem040 : Problem{
@@ -19,10 +17,10 @@ public class Problem040 : Problem{
     private int ff(int pos, int n) {
         int digits = 0;
         for (int x = 1; x < 1000000; x++) {
-            digits += ((int)Math.Log10(x)+1);
+            digits += (int)Math.Log10(x)+1;
             if (digits < pos) continue;
             string current = x.ToString();
-            return (int)current[current.Length - digits + pos - 1]-'0';
+            return current[current.Length - digits + pos - 1]-'0';
         }
         return -1;
     }
